♻️ refactor : 전환 로직 분리  #199
