+++ src/utils/editor/convert-dataurl-to-blob.ts
@@ -1,0 +1,13 @@
+export const dataURLToBlob = (dataUrl: string): Blob => {
+  const arr = dataUrl.split(',');
+  const mimeMatch = arr[0].match(/:(.*?);/);
+  if (!mimeMatch) throw new Error('유효하지 않은 Data URL');
+
+  const mime = mimeMatch[1];
+  const bstr = atob(arr[1]);
+  let n = bstr.length;
+  const u8arr = new Uint8Array(n);
+  while (n--) u8arr[n] = bstr.charCodeAt(n);
+
+  return new Blob([u8arr], { type: mime });
+};
